pyproject.toml
[tool.poetry]
name = "demand-forecasting-mlops"
version = "1.0.0"
description = "End-to-end MLOps platform for demand forecasting"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"

# Data Processing
pandas = "^2.1.0"
numpy = "^1.26.0"
pyarrow = ">=17.0.0"

# Data Validation
pandera = "^0.17.0"
great-expectations = "^0.18.0"
# ydata-profiling removed - not compatible with Python 3.13+

# ML Models
prophet = "^1.1.5"
xgboost = "^2.0.0"
lightgbm = "^4.1.0"
catboost = "^1.2.0"
scikit-learn = "^1.3.0"

# Hyperparameter Tuning
optuna = "^3.4.0"

# Experiment Tracking
mlflow = "^2.9.0"

# API
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Dashboard
streamlit = "^1.29.0"
plotly = "^5.18.0"

# Monitoring
evidently = "^0.4.0"

# Logging
loguru = "^0.7.0"

# Utilities
python-dotenv = "^1.0.0"
pyyaml = "^6.0.0"
joblib = "^1.3.0"
httpx = "^0.25.0"
dvc = "^3.66.1"
multimethod = "1.9.1"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"

# Code Quality
ruff = "^0.1.0"
black = "^23.0.0"
pre-commit = "^3.6.0"
mypy = "^1.7.0"

# Documentation
mkdocs = "^1.5.0"
mkdocs-material = "^9.5.0"

# Jupyter
jupyter = "^1.0.0"
ipykernel = "^6.27.0"

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]
target-version = "py310"

[tool.ruff.isort]
known-first-party = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
